[refactor(react)] 重構showFridgeAndRecipe
